templated the plugin manager.
most of the functions are now templates. still not all of them and i guess it will not work right now.
since it is a template the functions need to be instantiated once. usually this is done trough the calling code. the problem here is that the plugin manager should be a singletone class. therefore it needs to be instantiated once per new event type. to do this there is now a macro in the cc file which does exactly this. I am not sure if this is the most alegant way but it works